requirements.txt
django-extensions
django-jquery-ui
ufs-tools
django-compat
# optional
django-tastypie
django-bootstrap-form
django-tables2-reports
django-import-export==0.4.3  # 0.4.4 is not compatible with django-admin-bootstrapped, because admin/actions.html is not compatible
django-ajax-selects
django-reversion
django-guardian
django-mptt
django-crispy-forms